wmf: Rewrite update.sh based on npmjs.org tarball
There are two problems at the moment:

1. Source files and make artefacts from previous versions are
   left behind (due to git-merge).

2. Files that are meant to be excluded from the distribution
   (npmignore) are still kept, which might cause subtle
   differences. In particular, the Makefile installs an extra
   package (node_modules/marked) which is meant to be excluded.

I was half-way trying to make point 1 and 2 work by using a
combination of 'git reset $1', which looked as follows for the
first part of the update script:

  git reset "$1" .
  git add .gitreview README-Wikimedia update.sh
  git clean -dffx
  git add .
  git commit -m "Checkout upstream $1"

However that still didn't honour npmignore. It also still caused
one other minor side-effect, which is that due to running Make
ourselves, the dates in man files differ from the actual release.

This commit instead fetches and unpacks the tarball of the official
release. This is what npm uses when installing other packages,
and also when it does self-updates. It's also what upstreams
scary install.sh does.

Bug: T215562
Change-Id: I588a5f68d0508fca7223490e4d6a100d3c05febd